Move/expose testutils like UnixSocketServerThread
